requirements.txt
pyTelegramBotAPI==4.21.0
reportlab==4.1.0
psycopg[binary]>=3.1
psycopg_pool>=3.1
python-dotenv>=1.0.0
